cloud-run/ria-intelligence-api/requirements.txt
beautifulsoup4>=4.12.0,<5.0.0
fastapi>=0.115.0,<1.0.0
google-genai>=1.36.0,<2.0.0
google-cloud-storage>=2.19.0,<3.0.0
gunicorn>=23.0.0,<24.0.0
httpx>=0.28.0,<1.0.0
lxml>=5.0.0,<6.0.0
openai>=1.0.0,<2.0.0
Pillow>=12.1.1,<13.0.0
pydantic>=2.9.0,<3.0.0
uvicorn>=0.30.0,<1.0.0
